django/ReportBuilderDjangoTest/requirements.txt
astroid==2.2.5
certifi==2019.6.16
Django==2.2.3
django-report-builder==6.3.1
django-restframework==0.0.1
djangorestframework==3.10.1
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
openpyxl==2.6.2
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
